+++ backend/NobreLar/NobreLar/src/main/java/com/barueri/NobreLar/repository/CorretorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CorretorRepository extends JpaRepository<Corretor, Integer> {
+public interface CorretorRepository extends JpaRepository<Corretor, Long> {
 }
